requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-core==2.23.4
pyjwt==2.9.0
python-decouple==3.8
python-multipart==0.0.17
redis==5.2.0
sniffio==1.3.1
sqlalchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.2
typing-extensions==4.12.2
uvicorn==0.32.0
